requirements.txt
amqp==5.0.9
appdirs==1.4.3
APScheduler==3.8.1
apturl==0.5.2
asgiref==3.5.0
autopep8==1.6.0
Babel==2.9.1
backports.zoneinfo==0.2.1
bcrypt==3.1.7
billiard==3.6.4.0
blinker==1.4
Brlapi==0.7.0
cairocffi==1.3.0
CairoSVG==2.5.2
celery==5.2.3
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
command-not-found==0.3
crypto==1.4.1
cryptography==2.8
cssselect2==0.4.1
cupshelpers==1.0
cybersource-rest-client-python==0.0.33
DateTime==4.4
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
Django==2.2.27
django-celery-beat==2.2.1
django-environ==0.8.1
django-extra-views==0.13.0
django-haystack==3.1.1
django-json-tag==0.0.1
django-localflavor==3.1
django-oscar==3.1
django-oscar-paypal==2.0.0
django-phonenumber-field==5.0.0
django-tables2==2.3.4
django-timezone-field==4.2.3
django-treebeard==4.4
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-oauth==1.1.0
duplicity==0.8.12.0
entrypoints==0.3
factory-boy==3.2.1
Faker==13.0.0
fasteners==0.14.1
filelock==3.4.2
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
keyring==18.0.1
kombu==5.2.3
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
more-itertools==4.2.0
mysqlclient==2.1.0
Naked==0.1.31
netifaces==0.10.4
oauthlib==3.1.0
packaging==21.3
paramiko==2.6.0
pbr==5.4.5
pdfkit==1.0.0
pexpect==4.6.0
phonenumbers==8.12.42
Pillow==9.0.1
pipenv==2022.1.8
platformdirs==2.5.0
prompt-toolkit==3.0.26
purl==1.6
pycairo==1.16.2
pycodestyle==2.8.0
pycountry==22.1.10
pycparser==2.21
pycryptodome==3.14.1
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
PyNaCl==1.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyphen==0.12.0
python-apt==2.0.0+ubuntu0.20.4.7
python-crontab==2.6.0
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-stdnum==1.17
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.26
PyYAML==5.3.1
redis==4.1.2
requests==2.27.1
requests-unixsocket==0.2.0
-e git+https://github.com/CyberSource/cybersource-rest-samples-python.git@d92375fb1878ee810f4028a8850e97398533dbd0#egg=samples
SecretStorage==2.3.1
shellescape==3.8.1
simplejson==3.16.0
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.4.2
ssh-import-id==5.10
stevedore==1.32.0
systemd-python==234
tinycss2==1.1.1
toml==0.10.2
tzdata==2021.5
tzlocal==4.1
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
vine==5.0.0
virtualenv==20.0.17
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wadllib==1.3.3
wcwidth==0.2.5
WeasyPrint==52.5
webencodings==0.5.1
wrapt==1.13.3
xkit==0.0.0
zipp==1.0.0
zope.interface==5.4.0
